+++ docs/Changelog.md
@@ -11,6 +11,7 @@
 - Rename `MapDatabase` to `MapFile` [#1184](https://github.com/mapsforge/vtm/pull/1184)
   - Rename `MultiMapDatabase` to `MultiMapFile`
 - Update MapBox vector tile dependency [#1199](https://github.com/mapsforge/vtm/pull/1199)
+- Disable `vtm-android-gdx` [#1211](https://github.com/mapsforge/vtm/pull/1211)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.25.0)
 
@@ -32,7 +33,6 @@
 
 ## Version 0.23.0 (2025-02-22)
 
-- Android native libs for 16 KB page size [#1162](https://github.com/mapsforge/vtm/pull/1162)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.23.0)